fix: `forge stop` not working because of heart

+++ packages/forge-cli/src/cli/node/stop/stop.js
@@ -1,17 +1,28 @@
 const shell = require('shelljs');
 const { symbols, getSpinner } = require('core/ui');
-const { runNativeForgeCommand } = require('core/env');
+const { getForgeProcesses } = require('core/env');
 
-const stop = runNativeForgeCommand('stop', { silent: true });
-function main() {
-  shell.echo(`${symbols.success} Sending kill signal to forge daemon...`);
-  const spinner = getSpinner('Waiting for forge daemon to stop...');
-  spinner.start();
-  const { code, stderr } = stop();
-  if (code === 0) {
-    spinner.succeed('Forge daemon stopped!');
-  } else {
-    spinner.fail(`Forge daemon stop failed ${stderr}!`);
+async function main() {
+  try {
+    const list = await getForgeProcesses();
+    const heartProcess = list.find(x => x.name === 'heart');
+    if (!heartProcess) {
+      throw new Error('cannot get heart process info');
+    }
+
+    shell.echo(`${symbols.success} Sending kill signal to forge daemon...`);
+    const spinner = getSpinner('Waiting for forge daemon to stop...');
+    spinner.start();
+    const { code, stderr } = shell.exec(`kill ${heartProcess.pid}`);
+    if (code === 0) {
+      spinner.succeed('Forge daemon stopped!');
+    } else {
+      spinner.fail(`Forge daemon stop failed ${stderr}!`);
+    }
+    process.exit(0);
+  } catch (err) {
+    shell.echo(`${symbols.error} cannot get daemon process info, ensure forge is started!`);
+    process.exit(1);
   }
 }
